Ограничивает кол-во отзывов и мест неподалеку, выносит отзывы в отдельный компонент
